split inject for module (#36)
split button functions to inject.js
